Forntend 디렉토리, README.md 작성, Inputbar 수정, Favorites Page 수정
