+++ Frontend/src/assets/bear.png
