+++ Frontend/src/components/BookRecommendationItem.css
