+++ Frontend/src/components/BookRecommendationItem.tsx
