+++ Frontend/src/components/BookRecommender.css
@@ -18,7 +18,7 @@
 /* 채팅 패널 */
 .chat-panel {
   width: 1000px;
-  height: 80%;
+  height: 95%;
   margin: 0 auto;
   display: flex;
   flex-direction: column;
@@ -69,7 +69,7 @@
   align-items: center;
   padding: 1rem;
   background-color: #b3dff3;
-  margin-top: 10px;
+  margin-top: 2rem;
 }
 
 /* 입력창 */
